Move 'FlattenGroupIdentifiers' (and tests).
