--- internal/service/ec2/flex_test.go
@@ -1,29 +1,0 @@
-package ec2
-
-import (
-	"testing"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/ec2"
-)
-
-func TestFlattenGroupIdentifiers(t *testing.T) {
-	expanded := []*ec2.GroupIdentifier{
-		{GroupId: aws.String("sg-001")},
-		{GroupId: aws.String("sg-002")},
-	}
-
-	result := FlattenGroupIdentifiers(expanded)
-
-	if len(result) != 2 {
-		t.Fatalf("expected result had %d elements, but got %d", 2, len(result))
-	}
-
-	if result[0] != "sg-001" {
-		t.Fatalf("expected id to be sg-001, but was %s", result[0])
-	}
-
-	if result[1] != "sg-002" {
-		t.Fatalf("expected id to be sg-002, but was %s", result[1])
-	}
-}
