+++ internal/service/ec2/vpc_network_interface.go
@@ -1599,3 +1599,13 @@ func deleteLingeringComprehendENIs(ctx context.Context, g *multierror.Group, con
 
 	return nil
 }
+
+// Flattens security group identifiers into a []string, where the elements returned are the GroupIDs
+func FlattenGroupIdentifiers(dtos []*ec2.GroupIdentifier) []string {
+	ids := make([]string, 0, len(dtos))
+	for _, v := range dtos {
+		group_id := aws.StringValue(v.GroupId)
+		ids = append(ids, group_id)
+	}
+	return ids
+}
