+++ internal/service/ec2/vpc_network_interface_test.go
@@ -20,6 +20,27 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 )
 
+func TestFlattenGroupIdentifiers(t *testing.T) {
+	expanded := []*ec2.GroupIdentifier{
+		{GroupId: aws.String("sg-001")},
+		{GroupId: aws.String("sg-002")},
+	}
+
+	result := tfec2.FlattenGroupIdentifiers(expanded)
+
+	if len(result) != 2 {
+		t.Fatalf("expected result had %d elements, but got %d", 2, len(result))
+	}
+
+	if result[0] != "sg-001" {
+		t.Fatalf("expected id to be sg-001, but was %s", result[0])
+	}
+
+	if result[1] != "sg-002" {
+		t.Fatalf("expected id to be sg-002, but was %s", result[1])
+	}
+}
+
 func TestAccVPCNetworkInterface_basic(t *testing.T) {
 	var conf ec2.NetworkInterface
 	resourceName := "aws_network_interface.test"
